refactor: isolate frontend delivery from app bootstrap
